+++ src/commands/contract/compile.ts
@@ -2,7 +2,7 @@ import { Command, Flags } from "@oclif/core";
 import { spawn } from "node:child_process";
 import path = require("node:path");
 import { readdirSync } from "node:fs";
-import { ensureSwankyProject } from "../../lib/command-utils";
+import { ensureSwankyProject } from "../../lib/config";
 import { Spinner } from "../../lib/spinner";
 export class CompileContract extends Command {
   static description = "Compile the smart contract(s) in your contracts directory";
