+++ src/lib/command-utils.ts
@@ -1,6 +1,4 @@
 import execa = require("execa");
-import fs = require("fs-extra");
-import { SwankyConfig, DEFAULT_NETWORK_URL } from "../commands/init";
 
 export async function commandStdoutOrNull(
   command: string
@@ -10,36 +8,5 @@ export async function commandStdoutOrNull(
     return result.stdout;
   } catch {
     return null;
-  }
-}
-
-export async function ensureSwankyProject(): Promise<void> {
-  const configExists = await fs.pathExists("swanky.config.json");
-  if (!configExists) {
-    throw new Error("No 'swanky.config.json' detected in current folder!");
-  }
-}
-
-export async function getSwankyConfig(): Promise<SwankyConfig> {
-  try {
-    const config = await fs.readJSON("swanky.config.json");
-    return config;
-  } catch {
-    throw new Error("No 'swanky.config.json' detected in current folder!");
-  }
-}
-
-export function resolveNetworkUrl(
-  config: SwankyConfig,
-  networkName: string
-): string {
-  if (networkName === "") {
-    return DEFAULT_NETWORK_URL;
-  }
-
-  try {
-    return config.networks[networkName].url;
-  } catch {
-    throw new Error("Network name not found in SwankyConfig");
   }
 }
