+++ src/lib/template.ts
@@ -1,0 +1,18 @@
+import { readdirSync } from "fs-extra";
+import path = require("node:path");
+
+export function getTemplates(language = "ink") {
+  const templatesPath = path.resolve(__dirname, "..", "templates");
+  const contractTemplatesPath = path.resolve(templatesPath, "contracts", language);
+  const fileList = readdirSync(contractTemplatesPath, {
+    withFileTypes: true,
+  });
+  const contractTemplatesList = fileList
+    .filter((entry) => entry.isDirectory())
+    .map((entry) => ({
+      message: entry.name,
+      value: entry.name,
+    }));
+
+  return { templatesPath, contractTemplatesPath, contractTemplatesList };
+}
